+++ lib/app/bloc/base/bloc_message.dart
