+++ lib/app/bloc/base/common/common_bloc.dart
@@ -7,7 +7,7 @@ import 'package:meta/meta.dart';
 import '../../../../shared/exceptions/app_exception_wrapper.dart';
 import '../../../../shared/extensions/number_extension.dart';
 import '../base_bloc.dart';
-import 'bloc_message.dart';
+import '../bloc_message.dart';
 
 part 'common_event.dart';
 part 'common_state.dart';
