--- lib/domain/bloc/bloc_message.dart
@@ -1,11 +1,0 @@
-import 'package:equatable/equatable.dart';
-
-class BlocMessage extends Equatable {
-  final String content;
-  final DateTime? createAt;
-
-  BlocMessage(this.content) : createAt = DateTime.now();
-
-  @override
-  List<Object?> get props => [content, createAt];
-}
